requirements.txt
dnspython3
requests
torrent_parser
